feat: enhance deleteBreed function to check for existing breeds before deletion

+++ src/features/animals/cow/service/breedMethods/deleteBreed.ts
@@ -1,3 +1,5 @@
+import { CONFLICT, NOT_FOUND } from '../../../../../lib/constants/http';
+import AppError from '../../../../../lib/utils/AppError';
 import CowModel, { CowBreedModel } from '../../model/cow.model';
 
 import mongoose from 'mongoose';
@@ -7,21 +9,26 @@ export async function deleteBreed(breedId: string) {
   session.startTransaction();
 
   try {
+    const cowsWithBreed = await CowModel.find({ breed: breedId }).session(
+      session
+    );
+    if (cowsWithBreed.length > 0) {
+      throw new AppError(CONFLICT, `Existing cows with breed ${breedId}`);
+    }
+
     const deletedBreed = await CowBreedModel.findByIdAndDelete(breedId).session(
       session
     );
     if (!deletedBreed) {
       // If breed is not found, abort transaction and return null
-      await session.abortTransaction();
-      await session.endSession();
-      return null;
+      throw new AppError(NOT_FOUND, `Breed with id ${breedId} is not found`);
     }
 
     // updates every cow with deleted breed
-    await CowModel.updateMany(
-      { breed: breedId },
-      { $set: { breed: null } }
-    ).session(session);
+    // await CowModel.updateMany(
+    //   { breed: breedId },
+    //   { $set: { breed: null } }
+    // ).session(session);
 
     // confims transaction
     await session.commitTransaction();
